dom/base/test/test_messageChannel_unshipped.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=677638
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 677638 - unshipped message port queue</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=677638">Mozilla Bug 677638</a>
<div id="content"></div>
<pre id="test">
</pre>
  <script type="application/javascript">

  function test_orderedMessages() {
    var a = new MessageChannel();
    ok(a, "MessageChannel created");

    var b = new MessageChannel();
    ok(b, "MessageChannel created");

    var expectedNumber = 1;
    function testEvent(number, id) {
      is(expectedNumber, number, "This is the right number!");
      ok(!((expectedNumber - id) % 4), "From the right port: " + expectedNumber + " " + id);
      expectedNumber++;

      if (expectedNumber >100) {
        runTests();
      }
    }

    a.port1.onmessage = function(evt) {
      testEvent(evt.data, 2);
    };

    a.port2.onmessage = function(evt) {
      testEvent(evt.data, 1);
    };

    b.port1.onmessage = function(evt) {
      testEvent(evt.data, 4);
    };

    b.port2.onmessage = function(evt) {
      testEvent(evt.data, 3);
    };

    for (var i = 0; i < 100;) {
      a.port1.postMessage(++i);
      a.port2.postMessage(++i);
      b.port1.postMessage(++i);
      b.port2.postMessage(++i);
    }
  }

  function test_unstarted() {
    var a = new MessageChannel();
    ok(a, "MessageChannel created");

    var b = new MessageChannel();
    ok(b, "MessageChannel created");

    var expectedNumber = 1;
    function testEvent(number, id) {
      is(expectedNumber, number, "This is the right number!");
      ok(!((expectedNumber - id) % 3), "From the right port: " + expectedNumber + " " + id);
      expectedNumber++;

      // 102 because it's the first multiple of 3.
      if (expectedNumber > 102) {
        runTests();
      }
    }

    a.port1.onmessage = function(evt) {
      testEvent(evt.data, 2);
    };

    a.port2.onmessage = function(evt) {
      testEvent(evt.data, 1);
    };

    b.port1.addEventListener("message", function() {
      ok(false, "shouldn't be called");
    });

    b.port2.onmessage = function(evt) {
      testEvent(evt.data, 3);
    };

    for (var i = 0; i < 100;) {
      a.port1.postMessage(++i);
      a.port2.postMessage(++i);
      b.port1.postMessage(++i);
      b.port2.postMessage(1000);
    }
  }

  var tests = [
    test_orderedMessages,
    test_unstarted
  ];

  function runTests() {
    if (!tests.length) {
      SimpleTest.finish();
      return;
    }

    var test = tests.shift();
    test();
  }

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [["dom.messageChannel.enabled", true]]}, runTests);
  </script>
</body>
</html>
